File and stream regression fixes
git-svn-id: svn://svn.caucho.com/resin/trunk@1324 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/quercus/src/com/caucho/quercus/lib/file/BinaryInput.java
@@ -88,7 +88,7 @@ public interface BinaryInput extends BinaryStream {
   /**
    * Reads a line from the buffer.
    */
-  public StringValue readLine(int length)
+  public StringValue readLine(long length)
     throws IOException;
 
   /**
